requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cssselect2==0.6.0
Django==4.0.4
djangorestframework==3.13.1
drf-writable-nested==0.6.3
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.8.0
mysqlclient==2.1.0
oscrypto==1.3.0
Pillow==9.1.1
pisa==3.0.33
pdfrw==0.4
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
weasyprint==55.0
webencodings==0.5.1
xhtml2pdf @ git+https://github.com/chrisglass/xhtml2pdf.git@2d1ac88231ad1270817b7853777745f20b519240
pisa==3.0.33
python-dotenv==0.19.1
google-api-core==2.8.2
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
